requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anthropic==0.71.0
    # via -r requirements.in
anyio==4.11.0
    # via
    #   anthropic
    #   httpx
    #   watchfiles
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via posthog
bcrypt==5.0.0
    # via chromadb
build==1.3.0
    # via chromadb
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
charset-normalizer==3.4.4
    # via requests
chromadb==1.1.1
    # via -r requirements.in
click==8.3.0
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   tqdm
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
distro==1.9.0
    # via
    #   anthropic
    #   posthog
docstring-parser==0.17.0
    # via anthropic
durationpy==0.10
    # via kubernetes
filelock==3.20.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flatbuffers==25.9.23
    # via onnxruntime
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
google-auth==2.41.1
    # via kubernetes
googleapis-common-protos==1.70.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.75.1
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   anthropic
    #   chromadb
huggingface-hub==0.35.3
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
importlib-resources==6.5.2
    # via chromadb
jinja2==3.1.6
    # via torch
jiter==0.11.1
    # via anthropic
joblib==1.5.2
    # via scikit-learn
jsonschema==4.25.1
    # via chromadb
jsonschema-specifications==2025.9.1
    # via jsonschema
kubernetes==34.1.0
    # via chromadb
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via chromadb
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.4
    # via
    #   chromadb
    #   onnxruntime
    #   scikit-learn
    #   scipy
    #   transformers
oauthlib==3.3.1
    # via requests-oauthlib
onnxruntime==1.23.1
    # via chromadb
opentelemetry-api==1.38.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.38.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.38.0
    # via chromadb
opentelemetry-proto==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.38.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.59b0
    # via opentelemetry-sdk
orjson==3.11.3
    # via chromadb
overrides==7.7.0
    # via chromadb
packaging==25.0
    # via
    #   build
    #   huggingface-hub
    #   onnxruntime
    #   transformers
pillow==12.0.0
    # via sentence-transformers
posthog==5.4.0
    # via chromadb
protobuf==6.33.0
    # via
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pybase64==1.4.2
    # via chromadb
pydantic==2.12.3
    # via
    #   anthropic
    #   chromadb
pydantic-core==2.41.4
    # via pydantic
pygments==2.19.2
    # via rich
pypika==0.48.9
    # via chromadb
pyproject-hooks==1.2.0
    # via build
pyreadline3==3.5.4
    # via humanfriendly
python-dateutil==2.9.0.post0
    # via
    #   kubernetes
    #   posthog
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   transformers
    #   uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.9.18
    # via transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   kubernetes
    #   posthog
    #   requests-oauthlib
    #   transformers
requests-oauthlib==2.0.0
    # via kubernetes
rich==14.2.0
    # via
    #   chromadb
    #   typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
safetensors==0.6.2
    # via transformers
scikit-learn==1.7.2
    # via sentence-transformers
scipy==1.16.2
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.1.1
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   kubernetes
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
tenacity==9.1.2
    # via chromadb
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.1
    # via
    #   chromadb
    #   transformers
torch==2.9.0
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   chromadb
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.57.1
    # via sentence-transformers
typer==0.19.2
    # via chromadb
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   chromadb
    #   grpcio
    #   huggingface-hub
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sentence-transformers
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.3.0
    # via
    #   kubernetes
    #   requests
uvicorn[standard]==0.37.0
    # via chromadb
watchfiles==1.1.1
    # via uvicorn
websocket-client==1.9.0
    # via kubernetes
websockets==15.0.1
    # via uvicorn
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
